Refactor log filtering and add tabs (#167)
* Add a new DimensionModel interface

We'll implement this interface with other models to a model to handle its own sizing.

* Add a new sidebar layout

* Add a box layout

* Add tabs component

* Add wrapper to huh.Form

This wrapper implements the dimension model

* Prepare logs to live inside layout

Instead of having the logs view respond to the stack size, we will put it in a layout in a future commit. This change makes it implement the DimensionModel interface

* Return form fields when parsing cobra command

We return fields instead of a complete form so that we can define tabs with subsets of the fields.

This commit doesn't compile. The commit will fix up the logview to properly handle the tabs

* Move searching and footer to logview

* Fix up log test

* Add log view test

* Address comments

* Use space instead of x

* Fix form test