Change Redis product to Key Value  (#174)
Redis is now under the Render Key Value product name that encompasses both Redis and Valkey as available runtime engines.

This will move the JSON / YAML outputs under the key-value scoping, as well as add a new `kv-cli` command that replaces the existing `redis-cli` command and will use `redis-cli` or `valkey-cli` depending on what is available on the path.