[EXP-3146] Invalid timestamp message suggests a valid timestamp (#207)
`time.RFC3339` is not itself a valid timestamp, but only a valid format
string for timestamp parsing.

I've chosen to emit time.Now() in RFC-3339 rather than a fixed timestamp
because it will 1) emit in the user's timezone and 2) have an increased
chance of being closer to the time the user wants and therefore be
easier to edit.